requirements.txt
aiohttp>=3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
discord.py==2.2.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
requests>=2.28.2
soupsieve==2.4.1
urllib3==1.26.15
yarl==1.9.2
openai>=1.3.0
feedparser>=6.0.10
python-dateutil>=2.8.2
pytz>=2023.3
pytube>=15.0.0
markdownify>=0.11.6
google-api-python-client==2.108.0
beautifulsoup4>=4.12.2
google-genai>=0.3.0
